refactor, move files around delete dead code
