+++ Cargo.toml
@@ -14,21 +14,11 @@ tokio = { version = "1.40", features = ["full"] }
 tokio-util = "0.7.12"
 futures-util = "0.3.31"
 tracing = "0.1.40"
-slab = "0.4.9"
-rand_chacha = "0.3.1"
-parking_lot = "0.12.3"
-arc-metrics = "0.1"
 hotread = "0.1.1"
 serde = { version = "1", features = ["derive"] }
 
 [dev-dependencies]
 tracing-subscriber = "0.3"
-assert_matches = "1.5.0"
-clap = { version = "4.5", features = ["derive"] }
-rustyline = "14"
-axum = "0.7"
-tower-http = { version = "0.6", features = ["fs", "cors"] }
-serde_json = "1"
 ratatui = { version = "0.30.2", default-features = false, features = ["crossterm", "std"] }
 crossterm = "0.29.0"
 unicode-width = "0.2.2"
