--- src/cluster/election/mod.rs
@@ -1,14 +1,0 @@
-pub mod algorithm;
-pub mod paths;
-pub mod policy;
-pub mod timing;
-
-pub(crate) use algorithm::{
-    decide_election, ElectionDecision, ElectionInput, ElectionState, PeerReachability, TimedPeerObservation,
-};
-pub(crate) use paths::{append_path, valid_local_leader_path, valid_remote_leader_path};
-pub(crate) use policy::{
-    client::{observation_targets, ObservationTargetInput, PeerKind},
-    follow::{sort_follow_candidates, FollowCandidate},
-};
-pub use timing::NodeTiming;
