+++ src/lib.rs
@@ -1,13 +1,11 @@
 //! Synchronize deterministic state machines across a cluster.
 //!
-//! The crate is organized into focused layers:
-//! state management, protocol framing, transport adapters, cluster
-//! coordination, and test orchestration utilities.
+//! The crate is organized into focused layers: state management, protocol
+//! framing, transport adapters, and cluster coordination.
 
+pub mod cluster;
 pub mod protocol;
 pub mod state;
 pub mod transport;
-
-pub mod new;
 
 mod utils;
