+++ src/state/mod.rs
@@ -1,3 +1,5 @@
-pub mod determinstic_state;
+//! Deterministic state machines and the recovery/subscription layers on top.
+
+pub mod deterministic_state;
 pub mod recoverable_state;
-pub mod shared_state;
+pub mod subscribable_state;
